Update rect fields to handle signed values.
i3 defines these fields as uint32_t; Sway defines them as (signed) int. In
practice, it's possible to have negative x and y values if floating nodes are
partially offscreen, or with certain combinations of border settings and a
fullscreen window.

i64 should cover all the possibilities here: it's a strict superset of uint32_t
and either a superset of or the same as int, depending on the platform and
architecture.

+++ i3ipc-types/src/reply.rs
@@ -206,10 +206,10 @@ pub enum WindowType {
 
 #[derive(Deserialize, Serialize, Eq, PartialEq, Clone, Hash, Debug)]
 pub struct Rect {
-    pub x: usize,
-    pub y: usize,
-    pub width: usize,
-    pub height: usize,
+    pub x: i64,
+    pub y: i64,
+    pub width: i64,
+    pub height: i64,
 }
 
 #[derive(Deserialize, Serialize, Eq, PartialEq, Clone, Hash, Debug, Copy)]
